Moved the code behind codes to behavior class
